Add side effects procedure to make-mock
Add an optional argument to `make-mock` that generates side-effects,
instead of passing all calls to the mocked procedure. By default, this
procedure is the same as the mocked procedure, so this change should be
transparent to all existing client code.

This is useful in case the mocked procedure creates side effects that
have to be avoided during testing, e.g. connecting to a remote server,
writing to a database, removing files from filesystem ...