Change and fix define/mock
Fixes #17. Changes the define/mock grammar significantly.

- Mocks are now configured with a #:mock keyword and associated
keywords for the mock name and default value
- The mocked submodule can be specified explicitly with a #:in-submod
keyword
- define/mock-as was removed due to duplication

In the implementation, a define/mock-explicit form is added. This is
useful for macros that expand to a define/mock form. It’s not exposed
as I know of no external use cases, but it’s good to remember in case
one comes about.