Cleanup base implementation and make some API changes
1. Mock calls are no longer prefab, as it doesn’t make sense to make
them prefab when the values inside them might not be.
2. Keyword args are now stored in mock calls as an eq-based hash in a
separate field instead of a list of pairs.
3. The various mock-* procedures now consistently take the mock as the
first argument in keeping with the style pattern that “foo-x procedures
should take the foo they operate on as the first parameter”.
4. Minor readability issues like using `compose` and `thunk` where
applicable, defining a general `box-transform!` procedure, and moving
test imports up to the same location as normal imports.
5. Add Google to copyright holders.